Fix: sbd: not overwrite SYSCONFIG_SBD and sbd-disk-metadata if input 'n'(bsc#1194870)

+++ crmsh/bootstrap.py
@@ -357,6 +357,7 @@ Configure SBD:
         self.diskless_sbd = diskless_sbd
         self._sbd_devices = None
         self._watchdog_inst = None
+        self.no_overwrite = False
 
     def _parse_sbd_device(self):
         """
@@ -426,6 +427,7 @@ Configure SBD:
 
         configured_dev_list = self._get_sbd_device_from_config()
         if configured_dev_list and not confirm("SBD is already configured to use {} - overwrite?".format(';'.join(configured_dev_list))):
+            self.no_overwrite = True
             return configured_dev_list
 
         dev_list = []
@@ -472,6 +474,8 @@ Configure SBD:
         if self.diskless_sbd:
             return
         for dev in self._sbd_devices:
+            if self.no_overwrite and SBDManager.has_sbd_device_already_initialized(dev):
+                continue
             rc, _, err = invoke("sbd -d {} create".format(dev))
             if not rc:
                 error("Failed to initialize SBD device {}: {}".format(dev, err))
@@ -480,6 +484,10 @@ Configure SBD:
         """
         Update /etc/sysconfig/sbd
         """
+        if self.no_overwrite:
+            csync2_update(SYSCONFIG_SBD)
+            return
+
         shutil.copyfile(self.SYSCONFIG_SBD_TEMPLATE, SYSCONFIG_SBD)
         sbd_config_dict = {
                 "SBD_PACEMAKER": "yes",
@@ -518,9 +526,13 @@ Configure SBD:
         if not self._sbd_devices and not self.diskless_sbd:
             invoke("systemctl disable sbd.service")
             return
+        msg = ""
         if self.diskless_sbd:
             warn(self.DISKLESS_SBD_WARNING)
-        with status_long("Initializing {}SBD...".format("diskless " if self.diskless_sbd else "")):
+            msg = "Configuring diskless SBD"
+        elif not self.no_overwrite:
+            msg = "Initializing SBD"
+        with status_long(msg):
             self._initialize_sbd()
             self._update_configuration()
             invoke("systemctl enable sbd.service")
@@ -575,6 +587,15 @@ Configure SBD:
         if not dev_list:
             raise ValueError("No sbd device configured")
         inst._verify_sbd_device(dev_list, utils.list_cluster_nodes_except_me())
+
+    @staticmethod
+    def has_sbd_device_already_initialized(dev):
+        """
+        Check if sbd device already initialized
+        """
+        cmd = "sbd -d {} dump".format(dev)
+        rc, _, _ = utils.get_stdout_stderr(cmd)
+        return rc == 0
 
 
 _context = None
@@ -823,8 +844,9 @@ def status(msg):
 
 @contextmanager
 def status_long(msg):
-    log("# {}...".format(msg))
-    if not _context.quiet:
+    if msg:
+        log("# {}...".format(msg))
+    if not _context.quiet and msg:
         sys.stdout.write("  {}...".format(msg))
         sys.stdout.flush()
     try:
@@ -832,7 +854,8 @@ def status_long(msg):
     except:
         raise
     else:
-        status_done()
+        if msg:
+            status_done()
 
 
 def status_progress():
